Add files via upload
this one is a simple sudoku solver. the question board is already there in question u can edit the board to a sudoku board which u wanna enjoy and have great time cheating when ur friends are rying to solve their sudoku and complete in 2 mins tops. :)
Enjoy son, and if u loose in that sudoku comp. then broo don't come home again :(